+++ VinogradovAS/01_lab_polymons/include/node.h
@@ -7,4 +7,5 @@ struct TNode
 {
 	T value;
 	TNode<T>* pNext;
+	
 };
