+++ VinogradovAS/01_lab_polymons/samples/main.cpp
@@ -6,7 +6,7 @@
 #include <headlist.h>
 using namespace std;
 
-void ResultOutput(char op, TPolynom p1, TPolynom p2) {
+void ResultOutput(char op, TPolynom& p1, TPolynom& p2) {
 	if (op == '+') {
 		cout << "Результат: " << p1 + p2 << endl;
 	}
@@ -14,13 +14,22 @@ void ResultOutput(char op, TPolynom p1, TPolynom p2) {
 		cout << "Результат: " << p1 - p2 << endl;
 	}else cout << "Результат: " << p1 * p2 << endl;
 }
+int main1() {
+	string str = "x+x";
+	TPolynom p1(str);
+	TPolynom p2;
+	//p2.AddMonom(1, 1, 0, 0);
+	p2.AddMonom(1, 0, 0, 0);
+	cout << p1;
+	return 0;
+}
 
 int main()
 {
 	setlocale(LC_ALL, "Russian");
-	string str1, str2;
+	string str1 = "xyz", str2;
 	char op;
-	char flag = 1;
+	int flag = 1;
 	cout << "Калькулятор полиномов, который умеет их складывать, вычитать и умножать!" << endl << endl;
 	while (flag == 1) {
 		cout << "Введите первый полином:" << endl;
@@ -38,5 +47,6 @@ int main()
 		cin >> flag;
 		cout << endl;
 	}
+	
 	return 0;
 }
